rename clickhouse -> snuba
